# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelLLM gateway contracts",
    "version": "0.3.0",
    "description": "Phase 3 gateway: buffered generation, content/tool/schema enforcement, durable tenant-scoped operation metadata, optimistic versioned policy/registry updates, and Redis-backed admission. Public gateway and private inspection/MCP services have distinct boundaries."
  },
  "paths": {
    "/v1/chat/completions": {
      "post": {
        "summary": "Buffered text, approved tool proposals, and registered structured output",
        "security": [
          {
            "applicationBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/chat-request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approved result; preview may contain a denial decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/chat-response"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "413": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "429": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "502": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "503": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "504": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inspect": {
      "post": {
        "summary": "Phase 1 advisory preview, no execution",
        "security": [
          {
            "applicationBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/inspect-request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Approved result; preview may contain a denial decision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inspect-response"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "403": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "413": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "429": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "502": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "503": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "504": {
            "description": "Sanitized error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            },
            "headers": {
              "X-Sentinel-Operation-ID": {
                "description": "Server-generated operation UUID",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "applicationBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "serviceBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "chat-request": {
        "$ref": "schemas/chat-request.schema.json"
      },
      "chat-response": {
        "$ref": "schemas/chat-response.schema.json"
      },
      "inspect-request": {
        "$ref": "schemas/inspect-request.schema.json"
      },
      "inspect-response": {
        "$ref": "schemas/inspect-response.schema.json"
      },
...
      "error": {
        "$ref": "schemas/error.schema.json"
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development gateway"
    }
  ]
}